requirements.txt
antlr4-python3-runtime==4.7.2
bleach==3.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
docutils==0.15.2
future==0.18.2
idna==2.8
importlib-metadata==1.3.0
keyring==20.0.0
kiwisolver==1.1.0
matplot==0.1.9
matplotlib==3.1.2
more-itertools==8.0.2
networkx==2.4
numpy==1.17.4
pkginfo==1.5.0.1
pycparser==2.19
Pygments==2.5.2
pyloco==0.0.139
pyparsing==2.4.5
python-dateutil==2.8.1
python-telegram-bot==12.2.0
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
SimpleWebSocketServer==0.1.1
six==1.13.0
tornado==6.0.3
tqdm==4.40.2
twine==3.1.1
typing==3.7.4.1
urllib3==1.25.7
ushlex==0.99.1
webencodings==0.5.1
websocket-client==0.56.0
zipp==0.6.0
